% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dequeue.R
\name{dequeueJobs}
\alias{dequeueJobs}
\alias{dequeueDepends}
\title{Dequeue and run reverse-dependency checks, possibly in parallel}
\usage{
dequeueJobs(package, directory, exclude = NULL)

dequeueDepends(package, directory)
}
\arguments{
\item{package}{A character variable denoting a package}

\item{directory}{A character variable denoting a directory for the queuefile}

\item{exclude}{An optional character variable denoting an exclusion set csv file.}
}
\value{
A queue is create as a side effect, its elements are returned invisibly
}
\description{
Dequeue package for (potentially parallel) reverse-dependency check
}
\details{
This function consumes previously enqueued jobs for reverse dependency checks.
It is set up in such a way that multiple distinct and independent process can
run checks in parallel without effecting each other. If the underlying queue
file is on a network drive, this should may also work across multiple machines.
}
\author{
Dirk Eddelbuettel
}
