% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enqueue.R
\name{enqueueJobs}
\alias{enqueueJobs}
\alias{enqueueDepends}
\title{Enqueues reverse-dependent packages}
\usage{
enqueueJobs(package, directory)

enqueueDepends(package, directory)
}
\arguments{
\item{package}{A character variable denoting a package}

\item{directory}{A character variable denoting a directory}
}
\value{
A queue is create as a side effect, its elements are returned invisibly
}
\description{
Enqueue package for subsequent reverse-dependency check
}
\details{
This function relies on the \code{\link{available.packages}} function from R
along with the \code{liteq} package. The \code{getQueueFile} function is used to
determine the queue file directory and name.
}
\examples{
td <- tempdir()
options(repos=c(CRAN="https://cloud.r-project.org"))
jobsdf <- enqueueJobs(package="digest", directory=td)
}
\author{
Dirk Eddelbuettel
}
