% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getConnection}
\alias{getConnection}
\alias{createTable}
\alias{insertRow}
\title{Database Helper Functions}
\usage{
getConnection(file)

createTable(con)

insertRow(con, df)
}
\arguments{
\item{file}{A character variable pointing to a sqlite3 file}

\item{con}{A database connection object}

\item{df}{A one-row data.frame with results to be appended}
}
\value{
A connection object
}
\description{
Create a Database Connection
}
\author{
Dirk Eddelbuettel
}
