% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getDatabaseConnection}
\alias{getDatabaseConnection}
\alias{createTable}
\alias{insertRow}
\alias{createRunDataTable}
\title{Database Helper Functions}
\usage{
getDatabaseConnection(file)

createTable(con)

insertRow(con, df)

createRunDataTable(con)
}
\arguments{
\item{file}{A character variable pointing to a sqlite3 file}

\item{con}{A database connection object}

\item{df}{A one-row data.frame with results to be appended}
}
\value{
A connection object
}
\description{
Several Database Connection Helper Functions
}
\details{
These functions return a connection, create a table and insert
a row of data, respectively.
}
\author{
Dirk Eddelbuettel
}
