% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/utils.R
\name{runSanityChecks}
\alias{runSanityChecks}
\alias{getDataDirectory}
\alias{getQueueFile}
\alias{getConfig}
\alias{getExclusionSet}
\title{Various Helper Functions}
\usage{
runSanityChecks()

getDataDirectory(package, date = format(Sys.Date()), path = ".",
  sep = "_")

getQueueFile(package, date = format(Sys.Date()), path = ".", sep = "_")

getConfig(debug = FALSE)

getExclusionSet(exclusionfile)
}
\arguments{
\item{package}{Character variable providing the package name.}

\item{date}{Optional character variable describing a date, default
is current date.}

\item{path}{Option path, default is current directory.}

\item{sep}{Optional character with component separator, default is \dQuote{_}.}

\item{debug}{Optional boolean}
}
\value{
A directory name.
}
\description{
These functions create a data directory name given a package name
and additional optional arguments which can be used to store test
meta-data and results, return a queue file name, return the
configuration object (retrieved from several standard locations)
and load the exclusion set.
}
\details{
Additional helper functions get particular directories, config settings or
check some assumptions about files in \code{PATH} and alike,
}
\section{Config File}{

An optional config file can be used to set several configuration variables.
Currently supported are
\describe{
  \item{setup}{The path to an R file that will be \code{source}'ed.}
  \item{workdir}{The directory used for the parallel run of reverse depends.}
  \item{libdir}{The directory passed to \code{.libPaths} allow for additonal
build-dependencies.}
  \item{debug}{A boolean switch to enable more debugging output.}
  \item{verbose}{A boolean switch to enable more verbose output.}
}
The \code{setup} field is particularly useful as the file being sourced
can contain other R expressions setting environment variables or options.
}

\author{
Dirk Eddelbuettel
}
