% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDataDirectory}
\alias{getDataDirectory}
\alias{getQueueFile}
\title{Create a data directory}
\usage{
getDataDirectory(package, date = format(Sys.Date()), path = ".",
  sep = "_")

getQueueFile(package, date = format(Sys.Date()), path = ".", sep = "_")
}
\arguments{
\item{package}{Character variable providing the package name.}

\item{date}{Optional character variable describing a date, default
is current date.}

\item{path}{Option path, default is current directory.}

\item{sep}{Optional component separator, default is \dQuote{_}.}
}
\value{
A directory name.
}
\description{
This functions create a data directory given a package name and
additional optional arguments which can be used to store test
meta-data and results.
}
\author{
Dirk Eddelbuettel
}
